+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 🚀        IN PRODUCTION                        🚀
 (https://www.npmjs.com/package/@tiledesk/tiledesk-server/v/2.3.77) 
 
-# 2.3.44
+# 2.3.56
+- added telegram module
+- update tiledesk-apps to 0.1.17
+- update tiledesk-telegram-connector to 0.1.7
+
+# 2.3.55
 - updated tybot-connector to 0.1.97
 
 # 2.4.43